set up grid, add services, features
