+++ client/src/components/BusinessDetails/BusinessDetails.css
@@ -1,3 +1,23 @@
+.BusinessDetails-container {
+  text-align: center;
+  margin-top: 160px;
+  grid-column: 4/12;
+  display: flex;
+  flex-direction: row;
+  height: 50%;
+}
+
+.BusinessDetails-leftColumn {
+  width: 30%;
+  border: 0.5px solid black;
+  padding: 1.6rem;
+  text-align: left;
+}
+.BusinessDetails-rightColumn {
+  border: 0.5px solid black;
+  margin-left: 5%;
+  width: 65%;
+}
 .About {
   display: flex;
 }
